Move `run` to `cfg/eval/run.py` so the eval package init stays runtime-free
`cfg/eval/__init__.py` becomes a light relative-ref anchor (just `placeholder`); the runtime-dispatching
`run` config and its `inference` import move to a submodule. The task constants under `cfg.eval.real` now
import with no cycle and no lazy-import workaround.
